Cambios de Marketplace a Ecommers, pagos activados con shipify, cambios globales en aspecto del sitio web, versión movil y configuración para ecommers funcional

+++ src/app/navegacion/usuario/perfil/perfil/secciones/enviar-mensaje/componentes/mensaje/mensaje.component.ts
@@ -28,7 +28,7 @@ export class MensajeComponent implements OnInit{
       return date.getHours() + ':' + String(date.getMinutes()).padStart(2, '0');
     } else if (date.getFullYear() === now.getFullYear() && date.getMonth() === now.getMonth() && date.getDate() > now.getDate() - 7) {
       // Si la fecha es de esta semana, mostrar el día de la semana y la hora sin los segundos ni el indicador AM/PM
-      return date.toLocaleString('es-ES', { weekday: 'long' }) + ', ' + date.getHours() + ':' + String(date.getMinutes()).padStart(2, '0');
+      return date.toLocaleString('es-ES', { weekday: 'short' }) + ', ' + date.getHours() + ':' + String(date.getMinutes()).padStart(2, '0');
     } else {
       // Si la fecha no es de esta semana, mostrar la fecha completa sin la hora
       return date.toLocaleDateString();
